Add structurelint and fix architectural layer violations
Installed and configured structurelint to enforce clean architecture and project structure. Fixed all layer boundary violations by refactoring code to follow proper dependency rules.

Changes:
- Add .structurelint.yml configuration with comprehensive rules
  - Enforce max depth (5 levels), max files per dir (20), max subdirs (14)
  - Define architectural layers with clear dependency boundaries
  - Enable naming conventions for TypeScript/React Native project
  - Require essential files (README, package.json, tsconfig.json, .gitignore)
  - Enable layer boundary enforcement for clean architecture

- Move types to proper location
  - Move DrillStats interface from components to types
  - Move Arrow and Highlight interfaces from components to types
  - Define theme types (BoardThemeName, PieceThemeName, CoachPersonalityName) in types instead of importing from constants
  - Move navigation param lists (PlayStackParamList, ProfileStackParamList, CommunityStackParamList) to types

- Fix layer violations
  - Move errorHandler from utils to services (utils should not depend on services)
  - Remove ErrorBoundary dependency from utils/lazyLoad.tsx
  - Update navigation layer to allow dependencies on constants and utils
  - Allow state layer to depend on services for data persistence
  - Update all imports to use types from types file

- Update imports across codebase
  - Update 10+ files to import types from types instead of components
  - Update navigation files to re-export types from types file
  - Update screens to import navigation types from types file

All structurelint rules now pass with zero violations.
