+++ .structurelint.yml
@@ -1,0 +1,131 @@
+# structurelint configuration
+# Generated by: structurelint --init
+# Documentation: https://github.com/structurelint/structurelint
+
+root: true
+
+exclude:
+  - testdata/**
+  - .git/**
+  - node_modules/**
+  - dist/**
+  - build/**
+  - .expo/**
+  - .expo-shared/**
+  - android/**
+  - ios/**
+
+rules:
+  # Phase 0: Basic filesystem structure
+  max-depth:
+    max: 5
+
+  max-files-in-dir:
+    max: 20
+
+  max-subdirs:
+    max: 14
+
+  # Documentation requirements
+  file-existence:
+    required:
+      - path: README.md
+      - path: package.json
+      - path: tsconfig.json
+      - path: .gitignore
+
+  disallowed-patterns:
+    patterns:
+      - "*.tmp"
+      - "*.bak"
+      - ".DS_Store"
+      - "*.log"
+      - "*.swp"
+      - "*~"
+
+  # Naming conventions for React Native/TypeScript project
+  naming-convention:
+    rules:
+      # React components should be PascalCase
+      - pattern: PascalCase
+        paths:
+          - "src/components/**/*.tsx"
+          - "src/screens/**/*.tsx"
+          - "src/navigation/**/*.tsx"
+
+      # Services and utilities can be camelCase
+      - pattern: camelCase
+        paths:
+          - "src/services/**/*.ts"
+          - "src/utils/**/*.ts"
+          - "src/state/**/*.ts"
+
+      # Constants can be camelCase
+      - pattern: camelCase
+        paths:
+          - "src/constants/**/*.ts"
+
+  # Test file validation
+  test-location:
+    integration-test-dir: "tests"
+    allow-adjacent: true
+    exemptions:
+      - "testdata/**"
+
+  # Enforce layer boundaries for clean architecture
+  enforce-layer-boundaries:
+    enabled: true
+
+# Architectural layers for React Native app
+layers:
+  # Utilities and types - no dependencies
+  - name: utils
+    path: src/utils/**
+    dependsOn: []
+
+  - name: types
+    path: src/types/**
+    dependsOn: []
+
+  - name: constants
+    path: src/constants/**
+    dependsOn: []
+
+  # State management - can use types, constants, utils, and services for persistence
+  - name: state
+    path: src/state/**
+    dependsOn: [types, constants, utils, services]
+
+  # Services - business logic layer
+  - name: services
+    path: src/services/**
+    dependsOn: [types, constants, utils, state]
+
+  # Components - UI layer (can use services, state, types)
+  - name: components
+    path: src/components/**
+    dependsOn: [types, constants, utils, state, services]
+
+  # Screens - presentation layer (can use everything)
+  - name: screens
+    path: src/screens/**
+    dependsOn: [components, services, state, types, constants, utils]
+
+  # Navigation - can use screens, components, constants, and utils
+  - name: navigation
+    path: src/navigation/**
+    dependsOn: [screens, components, types, constants, utils]
+
+  # Root app - can depend on anything
+  - name: app
+    path: "App.tsx"
+    dependsOn: ["*"]
+
+# Dead code detection entrypoints
+entrypoints:
+  - index.ts
+  - App.tsx
+  - "**/*.test.ts"
+  - "**/*.test.tsx"
+  - "**/*.spec.ts"
+  - "**/*.spec.tsx"
